refactor(TopLevelNavigation): Move components
Move related components: masthead, sidebar,
notifications, search bar
