+++ packages/react-component-library/src/components/TopLevelNavigation/Masthead/Masthead.test.tsx
@@ -8,7 +8,7 @@ import {
   waitFor,
 } from '@testing-library/react'
 
-import { Link } from '../Link'
+import { Link } from '../../Link'
 import { Masthead } from './Masthead'
 import { Notification, Notifications } from '../NotificationPanel'
 import { MastheadUser } from './MastheadUser'
