+++ packages/react-component-library/src/components/TopLevelNavigation/Masthead/index.ts
