+++ packages/react-component-library/src/components/TopLevelNavigation/NotificationPanel/Notifications.test.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { render, RenderResult } from '@testing-library/react'
 
-import { Notification, Notifications } from '.'
-import { Link } from '../Link'
+import { Notification, Notifications } from './index'
+import { Link } from '../../Link'
 
 const MOCK_NOTIFICATION = (
   <Notification
