+++ packages/react-component-library/src/components/TopLevelNavigation/Sidebar/Sidebar.stories.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 
-import { CustomLink } from '../CustomLink'
-import { Graph, House, Tools } from '../../icons'
-import { Link } from '../Link'
+import { CustomLink } from '../../CustomLink'
+import { Graph, House, Tools } from '../../../icons'
+import { Link } from '../../Link'
 import { Notification, Notifications } from '../NotificationPanel'
-import { Sidebar, SidebarNav, SidebarNavItem, SidebarUser } from '.'
+import { Sidebar, SidebarNav, SidebarNavItem, SidebarUser } from './index'
 
 const stories = storiesOf('Sidebar', module)
