+++ packages/react-component-library/src/index.ts
@@ -28,17 +28,14 @@ export * from './components/Drawer'
 export * from './components/Dropdown'
 export * from './components/FormikGroup'
 export * from './components/Link'
-export * from './components/Masthead'
 export * from './components/Modal'
 export * from './components/Nav'
-export * from './components/NotificationPanel'
 export * from './components/NumberInput'
 export * from './components/Pagination'
 export * from './components/PhaseBanner'
 export * from './components/Popover'
 export * from './components/ProgressIndicator'
 export * from './components/Select'
-export * from './components/Sidebar'
 export * from './components/Switch'
 export * from './components/RangeSlider'
 export * from './components/Radio'
@@ -48,5 +45,8 @@ export * from './components/TabSet'
 export * from './components/TextArea'
 export * from './components/TextInput/TextInput'
 export * from './components/Tooltip'
+export * from './components/TopLevelNavigation/Masthead'
+export * from './components/TopLevelNavigation/NotificationPanel'
+export * from './components/TopLevelNavigation/Sidebar'
 export * from './components/Toast'
 export * from './components/Timeline'
